Fix portability edge-case
As reported by @knobo on GitHub, certain compilers will not return
an adjustable vector from #'DELETE, due to the ANSI specification
only requiring that the array element type remain unchanged, while
permitting the return value to be a simple vector; this can cause
a situation wherein the object's event silo is not extensible, and
an error will be signaled upon the next attempt to add a listener.

This commit provides a CONTINUE restart ensuring that the silo is
adjustable, along with a minor change to the portions of code that
are inlined; the change of inlining should not be noticeable.

The following line is metadata for GitHub:
close #4 #11